added tests for JsonWebKey extensions
